Reintroduce anchor detection as configurable step
Anchor detection got removed during refactoring, as a link will
still point to a sufficient url with a broken anchor, it still
generates wrong/broken links. Especially when markdown is used
to generate pages which are checked with tools like HTMLtest,
which do verify proper anchors etc.

Therefore i reintroduce this change with this pr, but as a
configurable option. This should still allow the proper
outcome of the refactoring, but ensures backwards compatibility
for other users who rely on proper anchor handling.

Relates: #24

Signed-off-by: Simon Schrottner <simon.schrottner@dynatrace.com>